+++ helperfunctions.h
@@ -7,24 +7,24 @@
 using namespace std;
 
 // Function to calculate a^b mod p without overflow
-long power_mod_p(long a, int b, int p);
+long power_mod_p(long a, long b, long p);
 
 //calculate x over p
-int legendre_symbol(int x, int p);
+int legendre_symbol(long x, long p);
 
 //find inverse of a mod p
-int inverse(int a, int p);
+long inverse(long a, long p);
 
 int squareroot(int a, int p);
 
 // Returs true if fourth power mod p, otherwise returns false
-bool isFourthPower(int x, int p);
+bool isFourthPower(long x, long p);
 
 //finds quartic residue classes
-vector<int> findQuarticResidueClasses(int p);
+vector<int> findQuarticResidueClasses(long p);
 
 // Function to generate prime numbers up to k
-vector<int> generate_primes_in_range(int lower, int upper);
+vector<long> generate_primes_in_range(long lower, long upper);
 
 int getAp(int p, int A, int B);
